Cleaned up redesign and shortened helper functions
