+++ src/app/page.tsx
@@ -2,14 +2,16 @@
 import Navbar from "@/components/Navbar/Navbar";
 import VideoPreview from "@/components/Video/Video";
 
-import { FaExclamationTriangle } from "react-icons/fa";
-
 import { useEffect, useState } from "react";
 import { REQUEST_VIDEO_API } from "../../private/constants";
 import VideoPlayer from "@/components/VideoPlayer/VideoPlayer";
 import DisplayName from "@/components/Displayname/Displayname";
-import getUserID from "@/helper/GetUserId";
+import { getUserID, makeNameReadable } from "@/helper/helperFunctions";
 import Divider from "@/components/Divider/Divider";
+import {
+  NoVideoFoundMain,
+  NoVideoFoundSide,
+} from "@/components/Novideofound/NoVideoFound";
 
 export interface VideoData {
   created_at: string;
@@ -96,18 +98,7 @@ export default function Home() {
               />
             </div>
           ) : (
-            <div className="h-full flex flex-col items-center justify-center">
-              <h2 className="flex items-center gap-2 text-lg font-semibold">
-                <FaExclamationTriangle className="text-yellow-500" />
-                No video available
-              </h2>
-              <p className="text-sm text-gray-500">
-                Please select a video from the list on the right
-              </p>
-              <p className="text-sm text-gray-500">
-                Or search for one with the search bar above
-              </p>
-            </div>
+            <NoVideoFoundMain />
           )}
         </section>
 
@@ -134,15 +125,7 @@ export default function Home() {
                 })}
               </div>
             ) : (
-              <div>
-                <p className="text-sm text-gray-500 bg-slate-200 border border-slate-300 p-3 rounded-lg my-2 flex items-center gap-2">
-                  <FaExclamationTriangle className="text-yellow-500" />
-                  No videos found for user
-                  <span className="text-slate-500 text-decoration-line: underline">
-                    {searchUser}
-                  </span>
-                </p>
-              </div>
+              <NoVideoFoundSide name={searchUser} />
             )}
           </div>
           <Divider />
@@ -167,12 +150,7 @@ export default function Home() {
                 })}
               </div>
             ) : (
-              <div>
-                <p className="text-sm text-gray-500 bg-slate-800 p-3 rounded-lg my-2 flex items-center gap-2">
-                  <FaExclamationTriangle className="text-yellow-500" />
-                  No videos found
-                </p>
-              </div>
+              <NoVideoFoundSide name={makeNameReadable(activeUser)} />
             )}
           </div>
         </section>
