--- src/helper/ConvertToEmbed.tsx
@@ -1,5 +1,0 @@
-const convertToEmbed = (url: string) => {
-  return url.replace("watch?v=", "embed/");
-};
-
-export default convertToEmbed;
