--- src/helper/GetUserId.tsx
@@ -1,5 +1,0 @@
-const getUserID = (word: string) => {
-  return word.toLowerCase().split(" ").join("_");
-};
-
-export default getUserID;
