--- src/helper/GetUsername.tsx
@@ -1,8 +1,0 @@
-const makeNameReadable = (name: string) => {
-  return name
-    ?.split("_")
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(" ");
-};
-
-export default makeNameReadable;
